moved various classes out of the Utilities namespace
