+++ src/Relatus/AssetManager.cs
@@ -1,6 +1,7 @@
 using Microsoft.Xna.Framework.Audio;
 using Microsoft.Xna.Framework.Content;
 using Microsoft.Xna.Framework.Graphics;
+using Relatus.Graphics;
 using Relatus.Utilities;
 
 namespace Relatus
