+++ src/Relatus/Utilities/Readers/DamPackerReader.cs
@@ -1,7 +1,5 @@
 using Microsoft.Xna.Framework.Content;
-using System;
-using System.Collections.Generic;
-using System.Text;
+using Relatus.Graphics;
 
 namespace Relatus.Utilities.Readers
 {
